refactor(split users_info): public API into user_info-sdk crate - Move public API traits, DTOs and errors into user_info-sdk - Keep domain service and local client implementation in user_info - Clarify ownership: sdk defines contracts, module owns behavior - Remove unnecessary wiring helpers for in-process clients
- docs
  - update module creation prompts and related documents for the latests changes in code
