+++ README.md
@@ -183,8 +183,7 @@ async fn list_resources_handler() -> Result<Json<Vec<MyResource>>, modkit::Probl
 ## Documentation
 
 - **[Module Development Guide](docs/MODKIT_UNIFIED_SYSTEM.md)** - How to create modules with the ModKit framework
-- **[Module Creation Prompt](docs/MODULE_CREATION_PROMT.md)** - Prompt for LLM-editor to generate a module from OpenAPI
-  specification
+- **[New Module Guideline](guidelines/NEW_MODULE.md)** - Step-by-step guide for creating production-grade modules
 - **[Contributing](CONTRIBUTING.md)** - Development workflow and coding standards
 
 ## Configuration
