--- examples/modkit/users_info/src/contract/mod.rs
@@ -1,7 +1,0 @@
-pub mod client;
-pub mod error;
-pub mod model;
-
-pub use client::*;
-pub use error::*;
-pub use model::*;
