--- examples/modkit/users_info/src/gateways/mod.rs
@@ -1,3 +1,0 @@
-pub mod local;
-
-pub use local::*;
