+++ examples/modkit/users_info/users_info/src/api/rest/sse_adapter.rs
