+++ examples/modkit/users_info/users_info/src/domain/events.rs
