+++ examples/modkit/users_info/users_info/src/infra/audit/http_audit_client.rs
