+++ examples/modkit/users_info/users_info/src/infra/storage/entity.rs
