+++ examples/modkit/users_info/users_info/src/infra/storage/odata_mapper.rs
